mysite/requirements.txt
django
django-cors-headers
django-rest-auth
djangorestframework
django-allauth
